Optimize goroutine's variable scope to avoid data race
Modified the scope of the variable 'node' within the goroutine in the range loop of controller.go. This was done to prevent data race conditions that could occur from concurrent reads/writes to the 'node' variable. The 'node' was moved into the function parameters of the goroutine, effectively confining its scope within the goroutine.