Add AWS credential support for node creation and version update
The main change in this commit is adding AWS credentials support to the process of node creation. This allows a session to be created using AccessKeyID and SecretAccessKey, which would help in creating an account, initializing a node and setting up the connection. Other dependencies, cloudflare, google and notify are also explicitly added to facilitate DDNS and enable notifications.

In addition, the version has been updated in version.go to mirror the new changes being made.

Finally, controller.go has been optimized by refactoring the common logic of node initialization into a new function buildNodes, in an effort to maintain cleaner and more intuitive code. Errors during the initialization process are immediately logged.