Update controller methods to public and refactor code
This commit makes several significant changes to the controller package. First, it updates the updateDomainIp() and isBlock() methods from the node.go file to be public functions by changing their names to UpdateDomainIp() and IsBlock(), providing better accessibility from other packages.

It also refactors the error handling part in the UpdateDomainIp() method by restructuring condition checks to reduce nesting and improve readability.

The diff also updates the Server struct to Service in the model.go file and makes corresponding changes in the controller.go file.

Finally, a new parameter timeout has been added to the node struct in the node_test.go file aiming to control the duration of the connection check.