Refactor node management logic and improve accessibility
This commit centralizes the node management logic from the controller package into a new dedicated app/node package. The purpose of this change is to improve the project's code organization and maintainability by separating concerns.

A Node model is introduced along with the refactors around the node management process. It moves several functions relating to nodes (like checkConnection, UpdateDomainIp, etc) into methods of the Node model, hence encapsulating the node-related operations. It also changes the visibility of some methods and struct fields for better accessibility from outside of the package.

Furthermore, renaming has been performed for clarity. For instance, the previously private Server struct in the model.go file is now renamed and public as Service.

Lastly, error handling in the UpdateDomainIp() function is refactored for better readability. This will help keep the project easier to understand and modify, enhancing its maintainability.