Remove dependency on ants concurrent pool
This commit removes dependency on 'github.com/panjf2000/ants/v2'. The ants worker pool was used previously due to concerns about executing too many goroutines concurrently. However, a manual worker channel with a max size has been initiated in the 'Service' struct as an alternative which simplifies the implementation. Additionally, the type of the worker channel has been changed from 'uint8' to 'bool' for better semantic clarity as it's utilized as a semaphore. It leads to removing unnecessary dependencies and reducing complexity, which makes the module easier to maintain.