Refactor Docker workflow and adjust version details
This commit simplifies the Docker workflow by dividing it into dev and release workflows. The Dockerfile is also updated to accept VERSION as a build arg. Similarly, version details have been calculated dynamically and renamed in version.go. This contributes to a more efficient and flexible workflow and better version tracking.