Refactor variables to public in Node struct
The variables in the Node struct of the node.go file have been refactored to public. The change involves renaming the variables and their references throughout the code from lowercase to uppercase. Furthermore, the setters and getter functions for these variables have been removed as they are unnecessary since the variables are now public.